+++ AppAsset/Audio.xcassets/Ballad of the Pegasus Cover.imageset/Contents.json
@@ -2,17 +2,17 @@
   "images" : [
     {
       "idiom" : "universal",
-      "filename" : "Image-187x187.png",
+      "filename" : "Image-283x283.png",
       "scale" : "1x"
     },
     {
       "idiom" : "universal",
-      "filename" : "Image-187x187@2x.png",
+      "filename" : "Image-283x283@2x.png",
       "scale" : "2x"
     },
     {
       "idiom" : "universal",
-      "filename" : "Image-187x187@3x.png",
+      "filename" : "Image-283x283@3x.png",
       "scale" : "3x"
     }
   ],
